+++ app/components/Menu/Menu.jsx
@@ -1,66 +1,108 @@
 import React, { Component } from 'react'
-import cx from 'classnames'
 import PropTypes from 'prop-types'
-import MenuItem from './MenuItem'
+import debounce from 'lodash/debounce'
+import MenuSheet from './MenuSheet'
+import noop from 'lodash/noop'
+import EventEmitter from 'eventemitter3'
 import MenuContextTypes from './MenuContextTypes'
-import { isFunction, isBoolean } from 'utils/is'
 
-
-function buildFilterIndex (items=[]) {
-  return items.map(({ name='' }) => name.toLowerCase())
+const keyCodes = {
+  left: 37,
+  up: 38,
+  right: 39,
+  down: 40,
+  esc: 27,
+  tab: 9,
+  enter: 13,
 }
 
-const MenuItemDivider = () => (<li><hr /></li>)
-
-class Menu extends Component {
-  constructor (props, context) {
+class MenuContainer extends Component {
+  constructor (props) {
     super(props)
-    this.state = {
-      activeItemIndex: this.getValidItemIndex(props.activeItemIndex, 'next', context)
+    this.state = { value: '' }
+    this.emitter = new EventEmitter()
+    this.reset = debounce(
+      () => this.setState({ value: '' })
+    , 500, { trailing: true })
+  }
+
+  getChildContext () {
+    const self = this
+    return {
+      subscribe (event, handler) {
+        self.emitter.on(event, handler)
+        return () => self.emitter.removeListener(event, handler)
+      },
+      setFocus (menuKey) {
+        self.setState({ currentFocus: menuKey })
+      },
+      getFocus () {
+        return self.state.currentFocus
+      },
+      menuContext: this.props.context,
+      deactivateTopLevelMenu: this.props.deactivate,
+      activatePrevTopLevelMenuItem: this.props.activatePrevTopLevelMenuItem,
+      activateNextTopLevelMenuItem: this.props.activateNextTopLevelMenuItem,
     }
-    this.itemNames2Index = buildFilterIndex(this.props.items)
-    this.menuItemInstances = []
   }
 
   componentWillMount () {
-    this.context.setFocus(this) // <- auto setFocus on mount
-    this.unsubscribe = this.context.subscribe('keyEvent', this.onKeyEvent)
+    this.deactivateTopLevelMenu = () => { this.props.deactivate() }
+    if (this.deactivateTopLevelMenu) window.addEventListener('click', this.deactivateTopLevelMenu)
+  }
+
+  componentDidMount () {
+    // we use tabindex to render the container a focus trap
+    // so that keyboard events can be listened in this scope
+    this.containerDOM.focus()
   }
 
   componentWillUnmount () {
-    if (this.unsubscribe) this.unsubscribe()
+    if (this.deactivateTopLevelMenu) window.removeEventListener('click', this.deactivateTopLevelMenu)
   }
 
-  getValidItemIndex = (index, direction='next', context) => {
-    const { menuContext } = this.context || context
-    const { items } = this.props
-    const item = items[index]
-    if (!item) return -1
-    const isDisabled = isBoolean(item.isDisabled) ? item.isDisabled
-      : isFunction(item.getIsDisabled) && item.getIsDisabled(menuContext)
-    const isHidden = isBoolean(item.isHidden) ? item.isHidden
-      : isFunction(item.getIsHidden) && item.getIsHidden(menuContext)
-    if (item.isDivider || isDisabled || isHidden) {
-      if (direction === 'next') {
-        return this.getValidItemIndex(index + 1, direction)
-      } else if (direction === 'prev') {
-        return this.getValidItemIndex(index - 1, direction)
-      }
-    } else {
-      return index
+  onKeyPress = (e) => {
+    const char = String.fromCharCode(e.charCode)
+    const value = this.state.value + char
+    this.setState({ value })
+    this.reset()
+    this.emitter.emit('keyEvent', { type: 'INPUT', value, target: this.state.currentFocus })
+  }
+
+  onKeyDown = (e) => {
+    const sendKeyEvent = eventType =>
+      this.emitter.emit('keyEvent', { type: eventType, target: this.state.currentFocus })
+    switch (e.keyCode) {
+      case keyCodes.up:
+        return sendKeyEvent('UP')
+      case keyCodes.down:
+        return sendKeyEvent('DOWN')
+      case keyCodes.left:
+        return sendKeyEvent('LEFT')
+      case keyCodes.right:
+        return sendKeyEvent('RIGHT')
+      case keyCodes.enter:
+        return sendKeyEvent('ENTER')
+
+      case keyCodes.esc:
+        this.props.deactivate()
+        return sendKeyEvent('ESC')
+
+      case keyCodes.tab:
+        e.preventDefault()
+        if (e.shiftKey) {
+          this.props.activatePrevTopLevelMenuItem()
+          return sendKeyEvent('SHIFT_TAB')
+        } else {
+          this.props.activateNextTopLevelMenuItem()
+          return sendKeyEvent('TAB')
+        }
+      default:
+        break
     }
-    return -1
+    e.stopPropagation()
   }
 
-  onMouseEnter = () => {
-    this.context.setFocus(this)
-  }
-
-  onMouseLeave = () => {
-    if (this.context.getFocus() === this) {
-      this.setState({ activeItemIndex: -2 })
-    }
-  }
 
   onFilterInputChange = (filterValue) => {
     const targetItemName = this.itemNames2Index
@@ -69,125 +111,43 @@ class Menu extends Component {
     this.activateItemAtIndex(targetIndex)
   }
 
-  onKeyEvent = (keyEvent) => {
-    // do not response if self is not the target
-    if (keyEvent.target !== this) return undefined
-    const currentIndex = this.state.activeItemIndex
-    const itemsLength = this.props.items.length
-    const activeItem = this.props.items[currentIndex]
-    const activeMenuItemInstance = this.menuItemInstances[currentIndex]
-    let nextIndex = -1
-    switch (keyEvent.type) {
-      case 'UP':
-        if (currentIndex < 0) {
-          nextIndex = itemsLength - 1
-        } else {
-          nextIndex = currentIndex - 1
-        }
-        if (nextIndex < 0) nextIndex = 0
-        nextIndex = this.getValidItemIndex(nextIndex, 'prev')
-        break
-
-      case 'DOWN':
-        if (currentIndex < 0) {
-          nextIndex = 0
-        } else {
-          nextIndex = currentIndex + 1
-        }
-        if (nextIndex >= itemsLength) nextIndex = itemsLength - 1
-        nextIndex = this.getValidItemIndex(nextIndex, 'next')
-        break
-
-      case 'LEFT':
-        if (this.props.isSubmenu) {
-          this.props.deactivate()
-        } else {
-          this.context.activatePrevTopLevelMenuItem()
-        }
-        break
-
-      case 'RIGHT':
-        if (activeItem && activeItem.items) {
-          if (activeMenuItemInstance) activeMenuItemInstance.showSubmenu()
-        } else {
-          this.context.activateNextTopLevelMenuItem()
-        }
-        break
-
-      case 'ENTER':
-        if (activeMenuItemInstance) activeMenuItemInstance.execCommand()
-        break
-
-      case 'INPUT':
-        this.onFilterInputChange(keyEvent.value)
-        break
-
-      default:
-    }
-    this.activateItemAtIndex(nextIndex)
-  }
-
-  activateItemAtIndex = (index) => {
-    if (index >= 0) this.setState({ activeItemIndex: index })
-  }
-
-  renderMenuItems (items) {
-    return items.map((item, i) => {
-      const key = `menu-item-${item.name}-${i}`
-
-      if (item.isDivider) {
-        return <MenuItemDivider key={key} />
-      }
-
-      if (isFunction(item.getIsHidden) && item.getIsHidden(this.context.menuContext)) {
-        return null
-      }
-
-      return (
-        <MenuItem item={item}
-          index={i}
-          key={key}
-          ref={r => this.menuItemInstances[i] = r}
-          isActive={this.state.activeItemIndex === i}
-          toggleActive={this.activateItemAtIndex}
-          parentMenu={this}
-        />
-      )
-    })
-  }
-
   render () {
-    const { items, className, style } = this.props
-
+    const { items, className, style, deactivate } = this.props
     return (
-      <ul className={cx('menu', className)}
-        style={style}
+      <div tabIndex='1'
+        ref={r => this.containerDOM = r}
         onClick={e => e.stopPropagation()}
-        onMouseEnter={this.onMouseEnter}
-        onMouseLeave={this.onMouseLeave}
+        onKeyDown={this.onKeyDown}
+        onKeyPress={this.onKeyPress}
       >
-        {this.renderMenuItems(items)}
-      </ul>
+        <MenuSheet items={items}
+          className={className}
+          style={style}
+          deactivate={deactivate}
+        />
+      </div>
     )
   }
 }
 
-Menu.propTypes = {
-  isSubmenu: PropTypes.bool.isRequired,
+MenuContainer.propTypes = {
   items: PropTypes.array.isRequired,
   deactivate: PropTypes.func.isRequired,
+  activateNextTopLevelMenuItem: PropTypes.func.isRequired,
+  activatePrevTopLevelMenuItem: PropTypes.func.isRequired,
+  context: PropTypes.any.isRequired,
   className: PropTypes.string,
   style: PropTypes.object,
-  activeItemIndex: PropTypes.number,
 }
 
-Menu.defaultProps = {
-  isSubmenu: false,
-  activeItemIndex: -1,
+MenuContainer.defaultProps = {
   className: '',
   style: {},
+  context: {},
+  activateNextTopLevelMenuItem: noop,
+  activatePrevTopLevelMenuItem: noop,
 }
 
-Menu.contextTypes = MenuContextTypes
+MenuContainer.childContextTypes = MenuContextTypes
 
-export default Menu
+export default MenuContainer
